Fix matt's suggestions

+++ hrpe/submit/create_submission.py
@@ -2,7 +2,7 @@ from zipfile import ZipFile
 from glob import glob
 from datetime import datetime
 from calendar import monthrange
-from os import remove
+import os
 from shutil import copyfile
 
 import pandas as pd
@@ -77,10 +77,11 @@ def _create_submission_csv(data, userinitials):
 def _zip_submissions_file(csv_path):
     # Zip the csv file defined the path and rename it to 'Predictions.zip'
     zip_name = "data/Predictions.zip"
+    if os.path.exists(zip_name):
+        os.remove(zip_name)
+
     req_name = "predictions.csv"
     req_path = os.path.join("data/submissions", req_name)
-
-    os.remove(zip_name)
 
     copyfile(csv_path, req_path)
